+++ openldap/openldap.d/local/share/cook/bin/configure-directories.sh
@@ -11,13 +11,6 @@ set -o pipefail
 
 export PATH=/usr/local/bin:$PATH
 
-# this is a bad setup
-# check that /mnt/openldap-data exists
-#if [ ! -d /mnt/openldap-data ]; then
-#    echo "ERROR: /mnt/openldap-data does not exist. Where is the persistent storage mount-in?"
-#    exit 1
-#fi
-
 # The database directory MUST exist prior to running slapd AND
 # should only be accessible by the slapd and slap tools.
 # Mode 700 recommended.